feat(ListaEnlazada): add error handling to 'eliminarDato' method
This commit adds error handling to the 'eliminarDato' method of the 'ListaEnlazada' class. If the list is empty, it throws a 'ListaEnlazadasExceptions' exception with the message "Lista vacia". If the data is not found in the list, it throws a 'ListaEnlazadasExceptions' exception with the message "Dato no encontrado". The method also now uses two pointers, 'actual' and 'anterior', to traverse the list and find the node that contains the data to be deleted. If the data is found, it is deleted, and the node count is decremented.