feat: Add new features to Interfaz class
Added the following methods to the Interfaz class:

regresar(): displays a message to prompt the user to press any key to return to the main menu, waits for a key press, and clears the console screen.
opcionInvalida(): displays a message to inform the user that an invalid option was selected, prompts them to press any key to continue, waits for a key press, and clears the console screen.
salir(): displays a farewell message to the user, prompts them to press any key to exit the program, waits for a key press, and closes the program